requirements_runtime.txt
aioredis==1.3.1
aiohttp==3.7.4
augly==0.1.6
boto3==1.17.17
cloudpickle==1.6.0
dill==0.3.6
foolbox==3.3.1
hydra-core==1.2.0
imagenet-c==0.0.3
mypy==0.812
nbconvert==6.2.0
numpy==1.21.2
pandas==1.3.4
pickle5==0.0.11
Pillow==8.3.2
protobuf==3.20.3
scikit-image==0.18.1
timm==0.4.12
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
webdataset==0.1.103
git+https://github.com/fra31/auto-attack@ad5c1ea36055caedbafb3326e4d6374658e8c33a
